+++ app/src/androidTest/java/co/touchlab/magicthreadsdemo/test/fromdb/FirstSaveTest.java
@@ -1,15 +1,14 @@
 package co.touchlab.magicthreadsdemo.test.fromdb;
 
 import android.os.Handler;
-import android.os.Looper;
-import android.test.ActivityInstrumentationTestCase2;
-import android.test.UiThreadTest;
+import android.test.ApplicationTestCase;
 
 import java.io.File;
 
 import co.touchlab.android.threading.tasks.persisted.PersistedTaskQueue;
 import co.touchlab.android.threading.tasks.persisted.storage.DefaultPersistedTaskQueue;
 import co.touchlab.android.threading.utils.UiThreadContext;
+import co.touchlab.magicthreadsdemo.DemoApplication;
 import co.touchlab.magicthreadsdemo.OptionsActivity;
 import co.touchlab.magicthreadsdemo.test.NetworkExceptionCommand;
 import co.touchlab.magicthreadsdemo.test.NeverCommand;
@@ -19,17 +18,17 @@ import co.touchlab.magicthreadsdemo.test.utils.ThreadHelper;
 /**
  * Created by kgalligan on 10/4/14.
  */
-public class FirstSaveTest extends ActivityInstrumentationTestCase2<OptionsActivity>
+public class FirstSaveTest extends ApplicationTestCase<DemoApplication>
 {
+    public FirstSaveTest()
+    {
+        super(DemoApplication.class);
+    }
+
     OptionsActivity activity;
     private Handler handler;
     PersistedTaskQueue queue;
     PersistedTaskQueue.PersistedTaskQueueState queueState;
-
-    public FirstSaveTest()
-    {
-        super(OptionsActivity.class);
-    }
 
     @Override
     protected void setUp() throws Exception
@@ -37,14 +36,16 @@ public class FirstSaveTest extends ActivityInstrumentationTestCase2<OptionsActiv
         super.setUp();
         handler = ThreadHelper.mainHandler();
 
-        File test = getInstrumentation().getTargetContext().getDatabasePath("test");
-        File[] files = test.getParentFile().listFiles();
-        for (File file : files)
+        File test = getContext().getDatabasePath("test");
+        File parentFile = test.getParentFile();
+        if(parentFile.exists())
         {
-            file.delete();
+            File[] files = parentFile.listFiles();
+            for (File file : files)
+            {
+                file.delete();
+            }
         }
-
-        activity = getActivity();
     }
 
     public void testInsertSavedValues()
@@ -56,7 +57,7 @@ public class FirstSaveTest extends ActivityInstrumentationTestCase2<OptionsActiv
             {
                 UiThreadContext.assertUiThread();
 
-                queue = DefaultPersistedTaskQueue.getInstance(getActivity());
+                queue = DefaultPersistedTaskQueue.getInstance(getContext());
                 queue.execute(new TestCommand());
                 queue.execute(new NetworkExceptionCommand());
                 queue.execute(new TestCommand());
@@ -70,7 +71,7 @@ public class FirstSaveTest extends ActivityInstrumentationTestCase2<OptionsActiv
             @Override
             public void run()
             {
-                queueState = queue.copyState();
+                queueState = queue.copyPersistedState();
             }
         }, 2200);
